Implement fix from Eric Roccasecca (via email)
Guards against an issue when starting the spinner multiple times.  It will now always kill the timer and make a new one when told to start animating, instead of starting more times (which cause the animation to run faster and faster).

Didn't apply the other fix (about the potential memory leak) because there is no -viewDidMoveToWindow method on CALayer.  Need to check if there is a corresponding method on CALayer to use for that purpose.

See [here](http://github.com/kelan/yrk-spinning-progress-indicator/commit/207c8b4c7cf4f6efd0d8de05fe74acc867f48e79) for the corresponding fix in the NSView-based project